+++ src/lib.rs
@@ -85,16 +85,15 @@ bag.finalize::<AlgorithmToUse>().await.unwrap();
 
 mod algorithm;
 mod checksum;
-#[cfg(feature = "generate")]
 mod generate;
 mod manifest;
+mod metadata;
 mod payload;
 mod read;
 
 /// Possible errors when manipulating BagIt containers
 pub mod error {
     pub use crate::checksum::ChecksumComputeError;
-    #[cfg(feature = "generate")]
     pub use crate::generate::GenerateError;
     pub use crate::payload::PayloadError;
     pub use crate::read::ReadError;
@@ -102,6 +101,7 @@ pub mod error {
 
 pub use algorithm::{Algorithm, ChecksumAlgorithm};
 pub use checksum::Checksum;
+use metadata::Metadata;
 pub use payload::Payload;
 
 #[derive(Debug, PartialEq)]
@@ -120,6 +120,9 @@ pub struct BagIt<'a, 'algo> {
 
     /// Which algorithm to use for checksums of the items
     checksum_algorithm: &'algo Algorithm,
+
+    /// Metadata tags
+    tags: Vec<Metadata>,
 }
 
 impl<'a, 'algo> BagIt<'a, 'algo> {
@@ -128,11 +131,13 @@ impl<'a, 'algo> BagIt<'a, 'algo> {
         directory: impl AsRef<std::path::Path>,
         items: Vec<Payload<'a>>,
         checksum_algorithm: &'algo Algorithm,
+        tags: Vec<Metadata>,
     ) -> Result<Self, error::ReadError> {
         Ok(Self {
             path: directory.as_ref().to_path_buf(),
             items,
             checksum_algorithm,
+            tags,
         })
     }
 
@@ -199,12 +204,10 @@ impl<'a, 'algo> BagIt<'a, 'algo> {
 
 #[cfg(test)]
 mod test {
-    use crate::{Algorithm, BagIt, Checksum, ChecksumAlgorithm, Payload};
+    use crate::{metadata::Metadata, Algorithm, BagIt, ChecksumAlgorithm, Payload};
     use sha2::Sha256;
-    use std::path::Path;
 
     #[tokio::test]
-    #[cfg(feature = "generate")]
     async fn generate_and_read_basic_bag_sha256() {
         let temp_directory = async_tempfile::TempDir::new().await.unwrap();
         let temp_directory = temp_directory.to_path_buf();
@@ -242,38 +245,37 @@ mod test {
             let expected = BagIt::from_existing_items(
                 temp_directory,
                 vec![
-                    Payload::new(
-                        Path::new("data/bagit.md"),
-                        Checksum::from(
-                            "eccdbbade12ba878af8f2140cb00c914f427405a987de2670e5c3014faf59f8e",
-                        ),
+                    Payload::test_payload(
+                        "data/bagit.md",
+                        "eccdbbade12ba878af8f2140cb00c914f427405a987de2670e5c3014faf59f8e",
+                        6302,
                     ),
-                    Payload::new(
-                        Path::new("data/paper_bag.jpg"),
-                        Checksum::from(
-                            "2b22a8fd0dc46cbdc7a67b6cf588a03a8dd6f8ea23ce0b02e921ca5d79930bb2",
-                        ),
+                    Payload::test_payload(
+                        "data/paper_bag.jpg",
+                        "2b22a8fd0dc46cbdc7a67b6cf588a03a8dd6f8ea23ce0b02e921ca5d79930bb2",
+                        19895,
                     ),
-                    Payload::new(
-                        Path::new("data/rfc8493.txt"),
-                        Checksum::from(
-                            "4964147d2e6e16442d4a6dbfbe68178a8f33c3e791c06d68a8b33f51ad821537",
-                        ),
+                    Payload::test_payload(
+                        "data/rfc8493.txt",
+                        "4964147d2e6e16442d4a6dbfbe68178a8f33c3e791c06d68a8b33f51ad821537",
+                        48783,
                     ),
-                    Payload::new(
-                        Path::new("data/sources.csv"),
-                        Checksum::from(
-                            "0fe3bd6e7c36aa2c979f3330037b220c5ca88ed0eabf16622202dc0b33c44e72",
-                        ),
+                    Payload::test_payload(
+                        "data/sources.csv",
+                        "0fe3bd6e7c36aa2c979f3330037b220c5ca88ed0eabf16622202dc0b33c44e72",
+                        369,
                     ),
-                    Payload::new(
-                        Path::new("data/totebag.jpg"),
-                        Checksum::from(
-                            "38ff57167d746859f6383e80eb84ec0dd84de2ab1ed126ad317e73fbf502fb31",
-                        ),
+                    Payload::test_payload(
+                        "data/totebag.jpg",
+                        "38ff57167d746859f6383e80eb84ec0dd84de2ab1ed126ad317e73fbf502fb31",
+                        10417,
                     ),
                 ],
                 algo.algorithm(),
+                vec![Metadata::PayloadOctetStreamSummary {
+                    octet_count: 85766,
+                    stream_count: 5,
+                }],
             )
             .unwrap();